fix: make union take float and int

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[int, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """
     Gets sum of a mixed list
     """
